Correcoes

+++ appdent/src/main/java/br/edu/infnet/appdent/CanalLoader.java
@@ -10,8 +10,9 @@ import org.springframework.boot.ApplicationRunner;
 import org.springframework.stereotype.Component;
 
 import br.edu.infnet.appdent.model.domain.Canal;
-import br.edu.infnet.appdent.model.domain.Paciente;
 import br.edu.infnet.appdent.model.service.CanalService;
+
+
 
 
 
@@ -38,7 +39,7 @@ public class CanalLoader implements ApplicationRunner{
 					
 					campos = linha.split(";");
 					
-					Canal canal = new Canal(campos[0], campos[1], campos[2]);
+					Canal canal = new Canal(campos[0], campos[1], Float.valueOf(campos[2]);
 					canal.setId(1);
 					canalService.incluir(canal);
